React Reorganization - wip rest of navigation
