+++ App/Containers/UserProfile/UserJourneys.js
@@ -13,7 +13,7 @@ import { removeJourney } from '../../Redux/StoriesRedux'
 
 // const UserJourneys = ({ myJourneys, myStoriesList, jid, name, current, screenProps}) => {
 class UserJourneys extends React.Component {
-  constructor(props) {
+  constructor (props) {
     super(props)
     this.state = {
       journeys: {}
@@ -22,19 +22,19 @@ class UserJourneys extends React.Component {
     this.pendingJourneys = new ListView.DataSource({rowHasChanged : (r1, r2) => r1 != r2})
     this.journeyRef = firebaseApp.database().ref('/users/' + this.uid + '/journeys/')
   }
-  componentDidMount() {this.listenForChange(this.journeyRef) }
+  componentDidMount () {this.listenForChange(this.journeyRef) }
   componentWillUnmount () {this.journeyRef.off('value', this.unsubscribe) }
 
-  listenForChange(ref) {
+  listenForChange (ref) {
     this.unsubscribe = ref.on('value', journeys => {
       let journeysObj = journeys.val()
       this.setState({ journeys: journeysObj })
     })
   }
 
-  listCurrentStory() {
+  listCurrentStory () {
     let userJourneys = this.state.journeys
-    if(userJourneys && userJourneys.current) {
+    if (userJourneys && userJourneys.current) {
       return <CurrentStoryInfo
         key={Object.keys(this.state.journeys.current)[0]}
         journey={Object.keys(this.state.journeys.current)[0]}
@@ -49,8 +49,8 @@ class UserJourneys extends React.Component {
     )
   }
 
-  listPendingInvites() {
-    if(this.state.journeys && this.state.journeys.pending){
+  listPendingInvites () {
+    if (this.state.journeys && this.state.journeys.pending){
       let pendList = this.pendingJourneys.cloneWithRows(Object.keys(this.state.journeys.pending))
       return (
         <ListView
@@ -68,7 +68,7 @@ class UserJourneys extends React.Component {
     )
   }
 
-  render() {
+  render () {
     const { myJourneys, myStoriesList, jid, name, current, screenProps} = this.props
     // journeys are not saved here but the story is, which has the story image
     // I think it would be nice to have the list of stories like the
